+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 - LASRVUCF - 29/04/2023 - Quit server file.
+- LACLI006 - 29/04/2023 - Fix close the streams of socket.
 
 ### Removed